Fix for #863 .
I referred to @jorisvandenbossche 's code very much.
I implemented three different na strategies (null, drop, keep).
Under my PC environment, the speed improvement was nearly nine times as follows.

%timeit list(df.iterfeatures())
Before: 3.36 s ± 225 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
After : 375 ms ± 14.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

(The same condition as @jorisvandenbossche 's [Notebook](http://nbviewer.jupyter.org/gist/jorisvandenbossche/3bea109e0622fa5e3f90eecdc496c816) )