@jorisvandenbossche 2018-12-09
@YuichiNotoya Thanks a lot for the PR!